pyproject.toml
[project]
name = "uab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.4",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pyside6>=6.10.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
